Move distancelabel to `base` module and remove now empty `vislab` module
