--- modules/vislab/include.cmake
@@ -1,1 +1,0 @@
-set(DEFAULT_MODULE ON)
